Keep Feed policy out of the budgeted view
The Claude Feed Bypass policy is shared testable behavior, but adding it directly to FeedPanelView pushed that already-budgeted file over the Swift length guard. Move the policy into a small Feed-owned source file and register it with the app target so the large view returns under its checked line budget.

Constraint: User requested continuing without another force-push

Constraint: CI failed scripts/swift_file_length_budget.py because FeedPanelView.swift exceeded its budget by two lines after review fixes

Rejected: Refresh the Swift file budget | the new lines are easy to split out and should not become accepted debt

Confidence: high

Scope-risk: narrow

Tested: python3 scripts/swift_file_length_budget.py --budget .github/swift-file-length-budget.tsv

Tested: plutil -lint GhosttyTabs.xcodeproj/project.pbxproj

Tested: git diff --check

Tested: python3 -m py_compile tests/test_claude_wrapper_hooks.py; bash -n Resources/bin/claude

Not-tested: Local XCTest execution per project policy
